backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
chardet==5.2.0
click==8.1.7
distro==1.9.0
fastapi==0.112.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
numpy==2.0.1
openai==1.40.2
pandas==2.2.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.5
